+++ CHANGELOG.md
@@ -1,11 +1,11 @@
 # Changelog
 ## v0.0.39 (in progress)
+- Create Object-by-Analyte EPIC directory schema 
 
 ## v0.0.38
 - Update Cell DIVE directory schema
 - Update Cell DIVE directory schema
 - Update legacy Cell DIVE directory schema
-- Create Object-by-Analyte EPIC directory schema 
 
 ## v0.0.37
 - Update Xenium directory schema