+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Changelog
 ## v0.0.39 (in progress)
+- Create Object-by-Analyte EPIC directory schema 
 
 ## v0.0.38
 - Update Cell DIVE directory schema