+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Update Xenium directory schema
 - Create Visium HD directory schema (provisional name)
 - Create Pixel-Seq v2 directory schema
+- Create MPLEx directory schema
 
 ## v0.0.38
 - Update Cell DIVE directory schema